requirements.txt
# Copyright (c) 2016-2025 Martin Donath <martin.donath@squidfunk.com>

# Permission is hereby granted, free of charge, to any person obtaining a copy
# of this software and associated documentation files (the "Software"), to
# deal in the Software without restriction, including without limitation the
# rights to use, copy, modify, merge, publish, distribute, sublicense, and/or
# sell copies of the Software, and to permit persons to whom the Software is
# furnished to do so, subject to the following conditions:

# The above copyright notice and this permission notice shall be included in
# all copies or substantial portions of the Software.

# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
# FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
# FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
# IN THE SOFTWARE.


pip install mkdocs-awesome-pages-plugin mkdocs-blog-plugin mkdocs-git-revision-date-localized-plugin mkdocs-redirects mkdocs-glightbox
pip install mkdocs-minify-plugin

# pip install mkdocs-macros-plugin       # For Jinja templating in Markdown
# pip install mkdocs-minify-plugin       # To minify HTML output
# pip install mkdocs-awesome-pages-plugin  # For auto-generating navigation
# pip install mkdocs-blog-plugin         # (Not needed if using Material's built-in blog feature)
# pip install mkdocs-git-revision-date-localized-plugin  # If you want Git commit dates in docs
# pip install mkdocs-redirects           # If handling redirects
# pip install mkdocs-glightbox            # If using image lightbox feature

# #### Jinja & Templating Dependencies
# pip install Jinja2 PyYAML markdown

# # Requirements for core
# jinja2~=3.1
# markdown~=3.2
# mkdocs~=1.6
# mkdocs-material-extensions~=1.3
# pygments~=2.16
# pymdown-extensions~=10.2

# Requirements for plugins
babel~=2.10
colorama~=0.4
paginate~=0.5
backrefs~=5.7.post1
requests~=2.26
babel==2.17.0
backrefs==5.8
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
csscompressor==0.9.5
ghp-import==2.1.0
htmlmin2==0.1.13
idna==3.10
Jinja2==3.1.6
jsmin==3.0.1
Markdown==3.7
MarkupSafe==3.0.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mkdocs-material==9.6.11
mkdocs-material-extensions==1.3.1
mkdocs-minify-plugin==0.8.0
packaging==24.2
paginate==0.5.7
pathspec==0.12.1
platformdirs==4.3.7
Pygments==2.19.1
pymdown-extensions==10.14.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
pyyaml_env_tag==0.1
requests==2.32.3
six==1.17.0
urllib3==2.3.0
watchdog==6.0.0